+++ src/phase.h
@@ -38,9 +38,10 @@ void simplify_global(tree_t top, hash_t *generics, jit_t *jit,
 // Perform static bounds checking
 void bounds_check(tree_t top);
 
-// Elaborate a top level design unit
-tree_t elab(object_t *top, jit_t *jit, unit_registry_t *ur, mir_context_t *mc,
-            cover_data_t *cover, sdf_file_t *sdf, rt_model_t *m);
+// Elaborate top level design units
+tree_t elab(object_t **top, int ntop, jit_t *jit, unit_registry_t *ur,
+            mir_context_t *mc, cover_data_t *cover, sdf_file_t *sdf,
+            rt_model_t *m);
 
 // Set the value of a top-level generic
 void elab_set_generic(const char *name, const char *value);